Fix Replace Random Generator
In file: fields.py, there is a method  that uses the pseudo-random number generator from {random} module. But the Python documentation suggests that the pseudo-random number generators of this module should not be used for security purposes. Instead, random number generators from the {secrets} module should be used. iCR replaced the usage of random number generators from the {random} module with the random number generator from the {secrets} module.